Day 17 part 2. I do not know why my solution works.
I just added more steps to the calculation, let it run for more initial
values of Vx and it just kept finding more solutions.

I clearly have not quite grasped a part of this problem and have
stumbled upon a way to find the correct answer.

I would like to come back and actually figure this out, but I'll do that
after I finish the rest of the problems.

For now, I'll take the star.